Moved start_log to utils.logging module and added SteamToLogger
