--- ci_lib/utils/__init__.py
@@ -1,1 +1,0 @@
-from .snakemake_tools import *
