+++ workflow/scripts/decoding.py
@@ -14,11 +14,12 @@ import sys
 sys.path.append(str((Path(__file__).parent.parent.parent).absolute()))
 
 from ci_lib.utils import snakemake_tools
+from ci_lib.utils.logging import start_log
 from ci_lib.features import Features, Means, Raws, Covariances, Correlations, AutoCovariances, AutoCorrelations, Moup
 
 #Setup
 # redirect std_out to log file
-logger = snakemake_tools.start_log(snakemake)
+logger = start_log(snakemake)
 if snakemake.config['limit_memory']:
     snakemake_tools.limit_memory(snakemake)
 try:
