+++ workflow/scripts/plot_performance.py
@@ -10,8 +10,9 @@ sys.path.append(str((Path(__file__).parent.parent.parent).absolute()))
 
 import ci_lib.plotting as plots
 from ci_lib.utils import snakemake_tools
+from ci_lib.utils.logging import start_log
 
-logger = snakemake_tools.start_log(snakemake)
+logger = start_log(snakemake)
 try:
     timer_start = snakemake_tools.start_timer()
