+++ workflow/scripts/run_aggregation.py
@@ -6,6 +6,7 @@ import sys
 sys.path.append(str((Path(__file__).parent.parent.parent).absolute()))
 
 from ci_lib.utils import snakemake_tools
+from ci_lib.utils.logging import start_log
 from ci_lib import DecompData
 import shutil
 import time
@@ -33,7 +34,7 @@ def rec_iter_link(input,output):
             print("error")
             raise TypeError('Input/Output structure missmatched')
 
-logger = snakemake_tools.start_log(snakemake)
+logger = start_log(snakemake)
 if snakemake.config['limit_memory']:
     snakemake_tools.limit_memory(snakemake)
 try:
